feat(ffi)!: delete the pre-v2 decode surface
Removes mfsk_decoder_new/_free, the MfskDecodeOptions handle and its
eight setters, mfsk_decode_i16/_f32, both sniper entry points,
mfsk_decode_i16_streaming, MfskResultList/MfskResult/
mfsk_result_list_free, and the MfskProtocol enum.

Deleted rather than deprecated because the two could not safely coexist:
both handles crossed as the same opaque MfskDecoder*, so a pointer that
wandered into the other's free function was UB no compiler could notice.
MfskProtocol was also the shape of the original problem -- one FST4
entry, so 15/30/120/300 were unreachable from C for decode and encode
alike.

WSPR/JT9/JT65 get their own entry points rather than a generic dispatch
that pretended they were the same shape as FT8. JT9 and JT65 are point
decodes at a known carrier, and the old path hardcoded 1500 and 1270 Hz
with no way to say otherwise; the frequency is an argument now. Q65
keeps its four functions and moves to the shared row type. Streaming
delivery moves onto the session via mfsk_session_set_on_decode, where it
reaches every mode instead of FT8 alone.

A session is single-threaded, where the old handle was not. That handle
carried one protocol tag so sharing it happened to work, and the driver
tested that it did; a session caches a callsign hash table it mutates on
every decode. This module's own docs already said a change adding cached
state must "tighten this documented contract back to strict
one-per-thread" -- this is that change.

Fixes a third instance of one class of defect, this one found by the C++
driver segfaulting: mfsk_mode_name((MfskMode)9999) crashed, because a
repr(C) fieldless enum is an int to C and reading an out-of-range
discriminant as a Rust enum lets the compiler assume it is one of the
listed variants. Modes, Q65 sub-modes and fading models now cross as
uint32_t and are validated; C callers still write MFSK_MODE_FT8, which
converts implicitly. The other two instances were the `&'static mut`
fabricated from a raw pointer and a memset options struct producing an
invalid MfskDecodeDepth. All three are "C hands Rust a value Rust's type
system says cannot exist", and all three are now checked at the boundary.

Test coverage is migrated, not dropped: the five FFI test files keep
what they measured (SIC recovering stations on qso3_busy, an AP hint
surfacing the blind-CQ decode, Q65's four strategies, the hash table
resolving a `<...>` callsign) against the new API, and the C++ driver
keeps every per-mode round trip plus the thread fan-out. One assertion
was corrected in the process: the AP test never claimed a decode-count
increase, and at these settings the plain path already reaches 14 on
that recording -- AP changes which passes earn them, not how many there
are.

mfsk.h loses 666 lines.

+++ mfsk-ffi/tests/mode_introspection.rs
@@ -99,7 +99,7 @@ fn names_round_trip_and_match_the_registry() {
         let mut m = MfskMode::Ft8;
         assert_eq!(unsafe { mfsk_mode_at(i, &mut m) }, MfskStatus::Ok);
 
-        let name = unsafe { CStr::from_ptr(mfsk_mode_name(m)) }
+        let name = unsafe { CStr::from_ptr(mfsk_mode_name(m as u32)) }
             .to_str()
             .unwrap();
         assert!(!name.is_empty());
@@ -152,7 +152,7 @@ fn a_typo_and_a_missing_feature_report_differently() {
 fn mode_info_reports_real_geometry() {
     let info = |m| {
         let mut i = std::mem::MaybeUninit::<MfskModeInfo>::zeroed();
-        let st = unsafe { mfsk_mode_info(m, i.as_mut_ptr()) };
+        let st = unsafe { mfsk_mode_info(m as u32, i.as_mut_ptr()) };
         assert_eq!(st, MfskStatus::Ok, "{m:?}");
         unsafe { i.assume_init() }
     };
@@ -200,7 +200,7 @@ fn mode_info_reports_real_geometry() {
 fn msk144_is_addressable_and_says_what_it_is_not() {
     let mut i = std::mem::MaybeUninit::<MfskModeInfo>::zeroed();
     assert_eq!(
-        unsafe { mfsk_mode_info(MfskMode::Msk144, i.as_mut_ptr()) },
+        unsafe { mfsk_mode_info(MfskMode::Msk144 as u32, i.as_mut_ptr()) },
         MfskStatus::Ok
     );
     let i = unsafe { i.assume_init() };
@@ -216,7 +216,7 @@ fn msk144_is_addressable_and_says_what_it_is_not() {
 /// including the two this branch just changed.
 #[test]
 fn published_capabilities_match_the_traits() {
-    assert_ne!(mfsk_mode_caps(MfskMode::Ft8) & MFSK_CAP_SNIPER, 0);
+    assert_ne!(mfsk_mode_caps(MfskMode::Ft8 as u32) & MFSK_CAP_SNIPER, 0);
     for m in [
         MfskMode::Ft4,
         MfskMode::Fst4s15,
@@ -224,17 +224,20 @@ fn published_capabilities_match_the_traits() {
         MfskMode::Fst4s300,
     ] {
         assert_eq!(
-            mfsk_mode_caps(m) & MFSK_CAP_SNIPER,
+            mfsk_mode_caps(m as u32) & MFSK_CAP_SNIPER,
             0,
             "{m:?} must not advertise a sniper it does not have"
         );
         assert_ne!(
-            mfsk_mode_caps(m) & MFSK_CAP_AP_WIDEBAND,
+            mfsk_mode_caps(m as u32) & MFSK_CAP_AP_WIDEBAND,
             0,
             "{m:?} does wide-band AP now and must say so"
         );
     }
-    assert_eq!(mfsk_mode_caps(MfskMode::Wspr) & MFSK_CAP_DECODE_HANDLE, 0);
+    assert_eq!(
+        mfsk_mode_caps(MfskMode::Wspr as u32) & MFSK_CAP_DECODE_HANDLE,
+        0
+    );
 }
 
 /// Defaults are data, published per mode — and carry the scale that
@@ -243,7 +246,7 @@ fn published_capabilities_match_the_traits() {
 fn defaults_are_published_with_their_scale() {
     let d = |m| {
         let mut x = std::mem::MaybeUninit::<MfskDecodeDefaults>::zeroed();
-        let st = unsafe { mfsk_mode_defaults(m, x.as_mut_ptr()) };
+        let st = unsafe { mfsk_mode_defaults(m as u32, x.as_mut_ptr()) };
         assert_eq!(st, MfskStatus::Ok, "{m:?}");
         unsafe { x.assume_init() }
     };
@@ -269,7 +272,7 @@ fn defaults_are_published_with_their_scale() {
 
     assert!(ft8.freq_max_hz > ft8.freq_min_hz && ft8.max_cand > 0);
     assert_eq!(
-        unsafe { mfsk_mode_defaults(MfskMode::Ft8, std::ptr::null_mut()) },
+        unsafe { mfsk_mode_defaults(MfskMode::Ft8 as u32, std::ptr::null_mut()) },
         MfskStatus::InvalidArg
     );
 }
@@ -285,7 +288,7 @@ fn size_versioning_writes_only_the_declared_prefix() {
     let mut buf = vec![0xAAu8; FULL];
     // An old header that knows only up to `ntones`.
     buf[..4].copy_from_slice(&(short as u32).to_ne_bytes());
-    let st = unsafe { mfsk_mode_info(MfskMode::Ft8, buf.as_mut_ptr() as *mut MfskModeInfo) };
+    let st = unsafe { mfsk_mode_info(MfskMode::Ft8 as u32, buf.as_mut_ptr() as *mut MfskModeInfo) };
     assert_eq!(st, MfskStatus::Ok);
 
     let written = u32::from_ne_bytes(buf[..4].try_into().unwrap()) as usize;
@@ -298,7 +301,7 @@ fn size_versioning_writes_only_the_declared_prefix() {
     // A zeroed struct means "same header as you", so everything is written.
     let mut full = std::mem::MaybeUninit::<MfskModeInfo>::zeroed();
     assert_eq!(
-        unsafe { mfsk_mode_info(MfskMode::Ft8, full.as_mut_ptr()) },
+        unsafe { mfsk_mode_info(MfskMode::Ft8 as u32, full.as_mut_ptr()) },
         MfskStatus::Ok
     );
     assert_eq!(unsafe { full.assume_init() }.size as usize, FULL);
